utils: move null_output_stream to utils
