+++ src/v/utils/null_output_stream.h
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2024 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+
+#include "base/seastarx.h"
+#include "base/units.h"
+
+#include <seastar/core/iostream.hh>
+#include <seastar/util/later.hh>
+
+namespace utils {
+// Data sink for noop output_stream instance
+// needed to implement scanning
+struct null_data_sink final : ss::data_sink_impl {
+    ss::future<> put(ss::net::packet data) final { return put(data.release()); }
+    ss::future<> put(std::vector<ss::temporary_buffer<char>> all) final {
+        return ss::do_with(
+          std::move(all), [this](std::vector<ss::temporary_buffer<char>>& all) {
+              return ss::do_for_each(
+                all, [this](ss::temporary_buffer<char>& buf) {
+                    return put(std::move(buf));
+                });
+          });
+    }
+    ss::future<> put(ss::temporary_buffer<char>) final { return ss::now(); }
+    ss::future<> flush() final { return ss::now(); }
+    ss::future<> close() final { return ss::now(); }
+};
+
+ss::output_stream<char> make_null_output_stream() {
+    auto ds = ss::data_sink(std::make_unique<null_data_sink>());
+    ss::output_stream<char> ostr(std::move(ds), 4_KiB);
+    return ostr;
+}
+
+} // namespace utils
